# api/openapi.yaml
openapi: 3.0.3
info:
  title: RebootFocus API
  version: 0.1.0
  description: Recovery coach API baseline for office workers
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/recovery/analytics/kpis/daily:
    post:
      tags: [Analytics]
      summary: Generate daily KPI mart
      operationId: generateDailyKpi
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateDailyKpiRequest"
      responses:
        "200":
          description: Daily KPI generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseDailyKpi"
    get:
      tags: [Analytics]
      summary: Get daily KPI mart
      operationId: getDailyKpi
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: string
        - in: query
          name: metricDate
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Daily KPI fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseDailyKpi"
  /api/v1/recovery/analytics/kpis/daily/quality:
    get:
      tags: [Analytics]
      summary: Get daily KPI quality report
      operationId: getDailyKpiQuality
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: string
        - in: query
          name: metricDate
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Daily KPI quality report fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseDailyKpiQuality"
  /api/v1/recovery/analytics/kpis/daily/backfill:
    post:
      tags: [Analytics]
      summary: Backfill daily KPI mart
      operationId: backfillDailyKpi
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BackfillDailyKpiRequest"
      responses:
        "200":
          description: Daily KPI backfill completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseBackfillDailyKpi"
  /api/v1/recovery/analytics/kpis/daily/last-processed-date:
    get:
      tags: [Analytics]
      summary: Get daily KPI last processed date
      operationId: getDailyKpiLastProcessedDate
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Daily KPI last processed date fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseDailyKpiLastProcessedDate"
...
components:
  schemas:
...
    ApiResponseDailyKpi:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: "#/components/schemas/DailyKpiResponse"
        message:
          type: string
          example: DAILY_KPI_FETCHED
        traceId:
          type: string
    ApiResponseDailyKpiQuality:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: "#/components/schemas/DailyKpiQualityResponse"
        message:
          type: string
          example: DAILY_KPI_QUALITY_FETCHED
        traceId:
          type: string
    ApiResponseBackfillDailyKpi:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: "#/components/schemas/BackfillDailyKpiResponse"
        message:
          type: string
          example: DAILY_KPI_BACKFILL_COMPLETED
        traceId:
          type: string
    ApiResponseDailyKpiLastProcessedDate:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: "#/components/schemas/DailyKpiLastProcessedDateResponse"
        message:
          type: string
          example: DAILY_KPI_LAST_PROCESSED_DATE_FETCHED
        traceId:
          type: string
...
    GenerateDailyKpiRequest:
      type: object
      required: [userId, metricDate]
      properties:
        userId:
          type: string
        metricDate:
          type: string
          format: date
    DailyKpiResponse:
      type: object
      properties:
        dailyKpiId:
          type: string
        userId:
          type: string
        metricDate:
          type: string
          format: date
        activation:
          type: boolean
        failureCount:
          type: integer
        recovery24:
          type: boolean
        recovery48:
          type: boolean
        restartCount24:
          type: integer
        restartCount48:
          type: integer
        ttrMinutes:
          type: integer
          format: int64
          nullable: true
        cycleCompletionRate:
          type: number
          format: double
        planExecutionRate:
          type: number
          format: double
        plannedWorkMinutes:
          type: integer
          format: int64
        actualWorkMinutes:
          type: integer
          format: int64
        estimationErrorMinutes:
          type: integer
          format: int64
        generatedAt:
          type: string
          format: date-time
    DailyKpiQualityResponse:
      type: object
      properties:
        dailyKpiQualityReportId:
          type: string
        userId:
          type: string
        metricDate:
          type: string
          format: date
        healthy:
          type: boolean
        duplicateRestartLinkCount:
          type: integer
        orphanRestartCount:
          type: integer
        restartBeforeFailureCount:
          type: integer
        lateRestartLinkCount:
          type: integer
        breakSessionReferenceCount:
          type: integer
        missingTimeboxReferenceCount:
          type: integer
        timezoneMismatchCount:
          type: integer
        totalIssueCount:
          type: integer
        generatedAt:
          type: string
          format: date-time
    BackfillDailyKpiRequest:
      type: object
      required: [userId, startDate, endDate]
      properties:
        userId:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
    BackfillDailyKpiResponse:
      type: object
      properties:
        userId:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        processedDays:
          type: integer
        processedMetricDates:
          type: array
          items:
            type: string
            format: date
        lastProcessedDate:
          $ref: "#/components/schemas/DailyKpiLastProcessedDateResponse"
    DailyKpiLastProcessedDateResponse:
      type: object
      properties:
        pipelineKey:
          type: string
        userId:
          type: string
        lastProcessedDate:
          type: string
          format: date
        updatedAt:
          type: string
          format: date-time
...